train.py
"""
Training Script
Trains RL agent on BTC/USD trading environment.
"""

import argparse
import yaml
from pathlib import Path


def load_config(config_path: str) -> dict:
    """
    Load configuration from YAML file.

    Args:
        config_path: Path to config file

    Returns:
        Configuration dictionary
    """
    # TODO: Implement in Epic 3.1
    pass


def train():
    """
    Main training function.
    """
    # TODO: Implement in Epic 3.1
    pass


if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Train RL agent for BTC/USD trading")
    parser.add_argument("--config", type=str, default="configs/training.yaml",
                       help="Path to training config file")
    args = parser.parse_args()

    train()
