requirements.txt
# Core RL and Gymnasium
gymnasium>=0.29.0
stable-baselines3>=2.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
PyYAML>=6.0

# Monitoring and Logging
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
